studies/caching/with-dagger-source/.dagger/pyproject.toml
[project]
name = "caching-dagger"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["dagger-io"]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
